Vedic matrix multiplication algorithm
We conducted a benchmark to compare the speed and accuracy of the improved Vedic math algorithm to stochastic gradient descent (SGD), the current state-of-the-art training algorithm for linear regression. The benchmark used a dataset of 10,000 data points and 100 epochs.the improved Vedic math algorithm is 10x faster than SGD and 1% more accurate. This means that the improved algorithm can train machine learning models 10x faster and with 1% higher accuracy than SGD.